backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
pydantic-settings==2.5.2
pydantic[email]==2.12.5
ollama==0.6.1
pypdf==4.3.1
python-docx==1.1.2
slowapi==0.1.9
python-dotenv==1.0.1
openpyxl==3.1.5
matplotlib==3.9.2
sentence-transformers==3.2.1
